% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{activate}
\alias{activate}
\alias{deactivate}
\title{Activate/Deactivate Sources}
\usage{
activate(.trj, source)

deactivate(.trj, source)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{source}{the name of the source or a function returning a name.}
}
\value{
Returns the trajectory object.
}
\description{
Activities for activating or deactivating the generation of arrivals by name.
}
\seealso{
\code{\link{set_trajectory}}, \code{\link{set_source}}.
}
