% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{activity.get_next}
\alias{activity.get_next}
\title{Get an activity's next activity}
\usage{
activity.get_next(activity)
}
\arguments{
\item{activity}{a pointer to the activity}
}
\description{
Get an activity's next activity
}

