% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_entities_with_interval}
\alias{add_entities_with_interval}
\title{Add multiple entities at once}
\usage{
add_entities_with_interval(sim_obj, n, name_prefix, trajectory_name, interval,
  start_time = 0)
}
\arguments{
\item{sim_obj}{the Simulator object}

\item{n}{the number of entities to add}

\item{name_prefix}{the name prefix to use for the new entitites}

\item{trajectory_name}{the name of the trajectory to use}

\item{interval}{the time between activations of each entity (is evaluated as an R command)}

\item{start_time}{the time at which the first entity is activated}
}
\description{
Add multiple entities at once
}

