% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_entity}
\alias{add_entity}
\title{Add an entity to the simulator object}
\usage{
add_entity(sim_obj, trajectory, name = "anonymous", activation_time = 0)
}
\arguments{
\item{sim_obj}{the simulator object}

\item{trajectory}{a trajectory object}

\item{name}{the name of the entity}

\item{activation_time}{the time at which to activate the entity}
}
\description{
Add an entity to the simulator object
}

