% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_event}
\alias{add_event}
\title{add an event to a trajectory}
\usage{
add_event(trajectory_obj, event_type, ...)
}
\arguments{
\item{event_type}{possible event types are \emph{seize}, \emph{release} and \emph{timeout}}

\item{resource_name}{(optional) used for specifying resource to be seized/released}

\item{resource_amount}{(optional) used for specifying amount of resource to be seized/released}

\item{time_value}{(optional) used for specifying timeout duration}

\item{event_id}{(optional) used to specify successor events, will auto generate if not specified, has to be unique}

\item{successor_id}{(optional) used to specify the successor of the current event, will automatically select next in line if not specified, \emph{"STOP"} means stop of trajectory}
}
\description{
add an event to a trajectory
}

