% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer-methods.R
\name{add_global}
\alias{add_global}
\title{Add a Global Attribute}
\usage{
add_global(.env, key, value)
}
\arguments{
\item{.env}{the simulation environment.}

\item{key}{the attribute name.}

\item{value}{the value to set, either a numeric or a \code{\link{schedule}},
so that the global may change during the simulation.}
}
\value{
Returns the simulation environment.
}
\description{
Attach a global variable to the simulation.
}
\seealso{
Convenience functions: \code{\link{schedule}}.
}
