% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_release_event}
\alias{add_release_event}
\title{add a release event to the trajectory object}
\usage{
add_release_event(trajectory_obj, resource, amount)
}
\arguments{
\item{trajectory_obj}{the trajectory object}

\item{resource}{the resource as a character string}

\item{amount}{the amount of resources to release}
}
\description{
add a release event to the trajectory object
}

