% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trajectory.R
\name{add_skip_event}
\alias{add_skip_event}
\title{add a skip event to the trajectory object}
\usage{
add_skip_event(trajectory_obj, number_to_skip)
}
\arguments{
\item{trajectory_obj}{the trajectory object}

\item{number_to_skip}{the number of events to skip following the skip event, will be evaluated seperately for each entity if a character is given (e.g. "sample(c(0,1),1)")}
}
\description{
add a skip event to the trajectory object
}

