% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_timeout_event}
\alias{add_timeout_event}
\title{add a time-out event to the trajectory object}
\usage{
add_timeout_event(trajectory_obj, duration)
}
\arguments{
\item{trajectory_obj}{the trajectory object}

\item{duration}{the duration of the time-out event}
}
\description{
add a time-out event to the trajectory object
}

