% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_trajectory}
\alias{add_trajectory}
\title{Add a trajectory}
\usage{
add_trajectory(sim_obj, name, trajectory_df)
}
\arguments{
\item{sim_obj}{the Simulator object}

\item{name}{the name of the trajectory}

\item{trajectory_df}{the trajectory data frame}
}
\description{
Add a trajectory
}

