% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_funcs.R
\name{at}
\alias{at}
\title{generate arrivals at specific times (generator convenience function)}
\usage{
at(times)
}
\arguments{
\item{times}{a (vector) of times at which to initiate an arrival}
}
\value{
returns a generator function
}
\description{
generate arrivals at specific times (generator convenience function)
}
\examples{
env <- simmer("SuperDuperSim") \%>\%
  add_resource("nurse", 1) \%>\%
  add_resource("doctor", 2) \%>\%
  add_resource("administration", 1) \%>\%
  add_generator("patient", t0, at(0,10,30))
}
\seealso{
\link{add_generator}
}

