% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_NNLS_functions.R
\name{build_kinetic_parameter_priors}
\alias{build_kinetic_parameter_priors}
\title{Build kinetic parameter priors}
\usage{
build_kinetic_parameter_priors(rxnSummary, kineticPars, omic_data)
}
\arguments{
\item{rxnSummary}{a list containing a reaction equation to fit (several forms of it) and multi-omic data}

\item{kineticPars}{table summarizing which species all non-linear kinetic parameters correspond to}

\item{omic_data}{abundances and uncertainties of reaction species and flux: setup in \code{\link{format_omic_data}}}
}
\value{
data.frame containing the parameters for each kinetic parameter's proposal distribution
}
\description{
Proposal distributions for each non-linear kinetic parameter are generated: these will be used by \code{\link{par_draw}}.
}

