% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QP_FBA_functions.R
\name{build_stoiMat}
\alias{build_stoiMat}
\title{Build stoichiometric matrix}
\usage{
build_stoiMat(metabolites, reactions, corrFile, rxnFile,
  internal_names = FALSE)
}
\arguments{
\item{metabolites}{metabolite IDs (e.g. s_0001)}

\item{reactions}{reaction IDs (e.g. r_0001)}

\item{corrFile}{mappings between metabolite IDs, names ...}

\item{rxnFile}{stoichometry of individual reactions}

\item{internal_names}{should row/column names be model IDs (internal_names = T) or common names (internal_names = F)}
}
\value{
a matrix with metabolites as rows, reactions as columns and reaction stoichiometry as values
}
\description{
Create the stoichiometry matrix, indicating the change of metabolites (rows)
per chemical reaction (columns)
}

