% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QP_FBA_functions.R
\name{calculate_QP_fluxes}
\alias{calculate_QP_fluxes}
\title{Calculate QP fluxes}
\usage{
calculate_QP_fluxes()
}
\value{
a data.frame containing all fluxes that were non-zero in at least one condition
}
\description{
Using boundary fluxes and reaction stoichiometry and directionality apply quadratic programming to
find a flux distribution that optimally agrees with boundary fluxes for each experimental condition.
}

