% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{clone}
\alias{clone}
\title{Add a clone activity}
\usage{
clone(traj, n, ...)
}
\arguments{
\item{traj}{the trajectory object.}

\item{n}{number of clones.}

\item{...}{optional sub-trajectories. Each clone will follow a different
sub-trajectory if available.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity that produces \code{n} clones of an arrival 
(the original one + \code{n-1} copies).
}
\seealso{
\code{\link{synchronize}}.
}

