% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_funcs.R
\name{every}
\alias{every}
\title{[Deprecated] Arrivals every specific interval}
\usage{
every(...)
}
\arguments{
\item{...}{a vector or multiple parameters of intervals between arrivals.}
}
\value{
Returns a generator function.
}
\description{
Generator convenience function to generate arrivals every specific interval.
When the generator reaches the last interval, it starts again from the first one.
}
\seealso{
\code{\link{add_generator}}.
}

