% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_NNLS_functions.R
\name{finalize_reaction_equations}
\alias{finalize_reaction_equations}
\title{Finalize reaction equations}
\usage{
finalize_reaction_equations(rxnEquations, all_species,
  kinetically_differing_isoenzymes)
}
\arguments{
\item{rxnEquations}{expressions for reaction equations}

\item{all_species}{all_species table summarizing the types of all reaction species}

\item{kinetically_differing_isoenzymes}{True-False: does only one equation exist for the reaction (false)
or are are reaction equations isoenzyme specific (true)}
}
\value{
expand rxnEquations to include kcat parameters and partial derivatives of rxnEquations w.r.t. each specie
}
\description{
Expression combining the log-occupancy equation and scaling of enzyme abundance by activity
determined w.r.t. metabolites/enzymes in log-space and linear-space
}
\details{
If there are multiple isoenzymes which differ kinetically then their expressions are generated seperately and then concatenated
a single equation is returned
}

