% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_NNLS_functions.R
\name{fit_reaction_equation_MCMC_NNLS}
\alias{fit_reaction_equation_MCMC_NNLS}
\title{Fit reaction equation using MCMC and NNLS}
\usage{
fit_reaction_equation_MCMC_NNLS(markov_pars, kineticPars, kineticParPrior,
  rxnEquations, all_species, omic_data, kinetically_differing_isoenzymes)
}
\arguments{
\item{markov_pars}{parameters for MCMC: a list containing 3 values: sample_freq, n_samples and burn_in}

\item{kineticPars}{table summarizing which species all non-linear kinetic parameters correspond to}

\item{kineticParPrior}{parameters for the proposal distribution of each kinetic parameter}

\item{rxnEquations}{expressions for reaction equations}

\item{all_species}{all_species table summarizing the types of all reaction species}

\item{omic_data}{abundances and uncertainties of reaction species and flux: setup in \code{\link{format_omic_data}}}

\item{kinetically_differing_isoenzymes}{True-False: does only one equation exist for the reaction (false)
or are are reaction equations isoenzyme specific (true)}
}
\value{
write draws from the posterior distribution of non-linear kinetic parameters (markov_par_vals)
and their log-likelihoods (lik_track) to the global environment
}
\description{
Run the MCMC-NNLS algorithm by continually updating kinetic parameters and evaluating their likelihood
}

