% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_NNLS_functions.R
\name{fit_reaction_equations}
\alias{fit_reaction_equations}
\title{Fit reaction equations}
\usage{
fit_reaction_equations(rxnSummary)
}
\arguments{
\item{rxnSummary}{a list containing a reaction equation to fit (several forms of it) and multi-omic data}
}
\value{
summary of MCMC run and input data
}
\description{
This function takes one reaction equation and associated multi-omic data and evaluate how well the
reaction equation can predict provided flux based on the MCMC-NNLS algorithm.
}

