% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QP_FBA_functions.R
\name{format_boundary_conditions}
\alias{format_boundary_conditions}
\title{Format boundary conditions}
\usage{
format_boundary_conditions()
}
\value{
write boundary flux inputs to global environment
}
\description{
Format experimental flux inputs and outputs for each condition (treatment_par) and match species to model reconstruction
}
\details{
load experimental data
combine boundary fluxes and uncertainties for each treatment into a list: treatment_par
flag the entries in the metabolic model that match boundary species
}

