% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_NNLS_functions.R
\name{format_omic_data}
\alias{format_omic_data}
\title{Format omic data}
\usage{
format_omic_data(kineticPars, all_species, rxnSummary,
  kinetically_differing_isoenzymes)
}
\arguments{
\item{kineticPars}{table summarizing which species all non-linear kinetic parameters correspond to}

\item{all_species}{all_species table summarizing the types of all reaction species}

\item{rxnSummary}{a list containing a reaction equation to fit (several forms of it) and multi-omic data}

\item{kinetically_differing_isoenzymes}{True-False: does only one equation exist for the reaction (false)
or are are reaction equations isoenzyme specific (true)}
}
\value{
a list summarizing the abundances and uncertainties in reaction metabolites, enzymes and flux
}
\description{
Align multi omic data across experimental conditions; Combine metabolites and enzymes
}

