% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{func_wrapper}
\alias{func_wrapper}
\title{Wrap a function / or value to be evaluated by the Rcpp backend}
\usage{
func_wrapper(value)
}
\arguments{
\item{value}{should be a value or a functuon}
}
\value{
a function which allows an attrs parameter
}
\description{
Wrap a function / or value to be evaluated by the Rcpp backend
}

