% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_capacity}
\alias{get_capacity}
\title{Get the capacity}
\usage{
get_capacity(env, name)
}
\arguments{
\item{env}{the simulation environment.}

\item{name}{the name of the resource.}
}
\value{
Returns a numeric value.
}
\description{
Gets a resource's capacity by name.
}
\seealso{
\code{\link{get_queue_size}}, \code{\link{set_capacity}}, \code{\link{set_queue_size}}, 
\code{\link{get_server_count}}, \code{\link{get_queue_count}}.
}

