% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/monitor.R
\name{get_entity_monitor_values}
\alias{get_entity_monitor_values}
\title{Get the monitor values of the entities in a simulation object}
\usage{
get_entity_monitor_values(sim_obj, aggregated = FALSE)
}
\arguments{
\item{sim_obj}{the simulation object}

\item{aggregated}{if FALSE (default) returns the raw data, otherwise returns an aggregated row per entity}
}
\description{
Get the monitor values of the entities in a simulation object
}

