% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{get_head}
\alias{get_head}
\title{Get the first activity}
\usage{
get_head(traj)
}
\arguments{
\item{traj}{the trajectory object.}
}
\value{
An external pointer to an activity object.
}
\description{
It can be used to get the pointer to a trajectory's first activity.
}
\seealso{
\code{\link{get_tail}}, \code{\link{get_n_activities}}, \code{\link{join}}.
}

