% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_mon_attributes}
\alias{get_mon_attributes}
\title{Get attribute statistics}
\usage{
get_mon_attributes(envs)
}
\arguments{
\item{envs}{the simulation environment (or a list of environments).}
}
\value{
Returns a data frame.
}
\description{
Gets the arrivals' attributes over time (if any).
}
\seealso{
\code{\link{get_mon_arrivals}}, \code{\link{get_mon_resources}}.
}

