% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_mon_resources}
\alias{get_mon_resources}
\title{Get resource statistics}
\usage{
get_mon_resources(envs, data = c("counts", "limits"))
}
\arguments{
\item{envs}{the simulation environment (or a list of environments).}

\item{data}{whether to retrieve the "counts", the "limits" or both.}
}
\value{
Returns a data frame.
}
\description{
Gets the resources' monitored data (if any).
}
\seealso{
\code{\link{get_mon_arrivals}}, \code{\link{get_mon_attributes}}.
}

