% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{get_n_activities}
\alias{get_n_activities}
\title{Get the number of activities}
\usage{
get_n_activities(traj)
}
\arguments{
\item{traj}{the trajectory object.}
}
\value{
The number of activities in the trajectory.
}
\description{
It can be used to get the total number of activities defined inside a trajectory.
}
\seealso{
Other methods to deal with trajectories:
\link{create_trajectory}, \link{show_trajectory}, \link{get_head},
\link{get_tail}, \link{seize}, \link{release}, \link{timeout}, 
\link{set_attribute}, \link{branch}, \link{rollback}.
}

