% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator-methods.R
\name{get_n_generated}
\alias{get_n_generated}
\alias{get_name}
\alias{get_attribute}
\title{Get process parameters}
\usage{
get_n_generated(.env, generator)

get_name(.env)

get_attribute(.env, key, global = FALSE)
}
\arguments{
\item{.env}{the simulation environment.}

\item{generator}{the name of the generator.}

\item{key}{the attribute name.}

\item{global}{if \code{TRUE}, the attribute will be global instead of per-arrival.}
}
\description{
Getters for processes (generators and arrivals) number of arrivals generated by a generator, the name
of the active arrival, an attribute from the active arrival or a global one.
}
\details{
\code{get_n_generated} returns the number of arrivals generated by a
given generator. \code{get_name} returns the number of the running arrival,
therefore is meant to be used inside a trajectory; otherwise, there will be
no arrival running and the function will throw an error. \code{get_attribute}
returns a running arrival's attribute or a global one. It is also meant to be
used inside trajectories. If the provided key was not previously set, it
returns a missing value.
}
\seealso{
\code{\link{set_attribute}}.
}
