% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity-methods.R
\name{get_next_activity}
\alias{get_next_activity}
\alias{get_prev_activity}
\title{Get the next/prev activity}
\usage{
get_next_activity(activity)

get_prev_activity(activity)
}
\arguments{
\item{activity}{an external pointer to the activity.}
}
\value{
Returns an external pointer to an activity object.
}
\description{
It takes an external pointer to an activity an returns the next/prev activity.
}
\seealso{
\code{\link{head.trajectory}}, \code{\link{tail.trajectory}}, \code{\link{print.externalptr}}.
}

