% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_queue_count}
\alias{get_queue_count}
\title{Get the queue count}
\usage{
get_queue_count(env, name)
}
\arguments{
\item{env}{the simulation environment.}

\item{name}{the name of the resource.}
}
\value{
Returns a numeric value.
}
\description{
Gets the number of customers in a resource's queue by name.
}
\seealso{
Other methods to deal with a simulation environment:
\link{simmer}, \link{now}, \link{peek}, \link{peek}, \link{onestep}, 
\link{run}, \link{add_resource}, \link{add_generator}, \link{get_mon_arrivals}, 
\link{get_mon_resources}, \link{get_capacity}, \link{get_queue_size}, \link{get_server_count}.
}

