% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_resource_monitor_values}
\alias{get_resource_monitor_values}
\title{Get the monitor values of a specified resource}
\usage{
get_resource_monitor_values(sim_obj, resource_name)
}
\arguments{
\item{sim_obj}{the simulation object}

\item{resource_name}{the name of the resource}
}
\description{
Get the monitor values of a specified resource
}

