% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/monitor.R
\name{get_resource_system_mon_values}
\alias{get_resource_system_mon_values}
\title{Get the system-wide monitor values (queue+server) of a specified resource}
\usage{
get_resource_system_mon_values(sim_obj, resource_name)
}
\arguments{
\item{sim_obj}{the simulation object}

\item{resource_name}{the name of the resource}
}
\description{
Get the system-wide monitor values (queue+server) of a specified resource
}

