% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_server_count}
\alias{get_server_count}
\title{Get the server count}
\usage{
get_server_count(env, name)
}
\arguments{
\item{env}{the simulation environment.}

\item{name}{the name of the resource.}
}
\value{
Returns a numeric value.
}
\description{
Gets the number of customers in a resource's server by name.
}
\seealso{
\code{\link{get_capacity}}, \code{\link{get_queue_size}}, \code{\link{set_capacity}}, 
\code{\link{set_queue_size}}, \code{\link{get_queue_count}}.
}

