% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer-methods.R
\name{get_sources}
\alias{get_sources}
\alias{get_resources}
\title{Get Sources and Resources Defined}
\usage{
get_sources(.env)

get_resources(.env)
}
\arguments{
\item{.env}{the simulation environment.}
}
\value{
A character vector.
}
\description{
Get a list of names of sources or resources defined in a simulation environment.
}
