% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{get_tail}
\alias{get_tail}
\title{Get the last activity}
\usage{
get_tail(traj)
}
\arguments{
\item{traj}{the trajectory object.}
}
\value{
An external pointer to an activity object.
}
\description{
It can be used to get the pointer to a trajectory's last activity.
}
\seealso{
\code{\link{get_head}}, \code{\link{get_n_activities}}, \code{\link{join}}.
}

