% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{handle_unfinished}
\alias{handle_unfinished}
\title{Handle Unfinished Arrivals}
\usage{
handle_unfinished(.trj, handler)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{handler}{trajectory object to handle unfinished arrivals. A \code{NULL}
value will unset the drop-out trajectory.}
}
\value{
Returns the trajectory object.
}
\description{
Activity for setting a drop-out trajectory for unfinished arrivals, i.e.,
those dropped from a resource (due to preemption, resource shrinkage or a
rejected \code{\link{seize}}) or those that \code{\link{leave}} a trajectory.
}
\seealso{
\code{\link{leave}}, \code{\link{set_capacity}}
}
