% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{head.trajectory}
\alias{get_head}
\alias{get_tail}
\alias{head.trajectory}
\alias{tail.trajectory}
\title{Return the first/last activity}
\usage{
\method{head}{trajectory}(x, ...)

get_head(.trj)

\method{tail}{trajectory}(x, ...)

get_tail(.trj)
}
\arguments{
\item{x}{the trajectory object.}

\item{...}{arguments to be passed to or from other methods.}

\item{.trj}{the trajectory object.}
}
\value{
Returns an external pointer to an activity object.
}
\description{
Trajectory getters for obtaining the pointer to its first/last activity.
}
\seealso{
\code{\link{get_n_activities}}, \code{\link{join}}.
}

