% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{join}
\alias{join}
\title{Join trajectories}
\usage{
join(...)
}
\arguments{
\item{...}{trajectory objects.}
}
\value{
A new trajectory object.
}
\description{
Concatenate any number of trajectories in the order specified.
}
\seealso{
Other methods for dealing with trajectories:
\link{create_trajectory}, \link{get_head}, \link{get_tail},
\link{get_n_activities}, \link{seize}, \link{release}, \link{timeout}, 
\link{set_attribute}, \link{branch}, \link{rollback}.
}

