% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{KeyValueMonitor-class}
\alias{KeyValueMonitor}
\alias{KeyValueMonitor-class}
\title{KeyValueMonitor class.}
\description{
Holds data not related to specific point in simulation time. E.g. holds values for keys suchs as "flow_time", "activity_time". NOT USED AT THE MOMENT
}

