% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{leave}
\alias{leave}
\title{Leave the Trajectory}
\usage{
leave(.trj, prob)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{prob}{a probability or a function returning a probability.}
}
\value{
Returns the trajectory object.
}
\description{
Activity for leaving the trajectory with some probability.
}
