% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QP_FBA_functions.R
\name{load_metabolic_model}
\alias{load_metabolic_model}
\title{Load metabolic model}
\usage{
load_metabolic_model()
}
\value{
assign metabolic model files to global environment
}
\description{
Load core model derived from SBML reconstruction
}

