% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{log_}
\alias{log_}
\title{Add a logging activity}
\usage{
log_(.trj, message)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{message}{the message to display, accepts either a string or a callable object
(a function) which must return a string.}
}
\value{
Returns the trajectory object.
}
\description{
Display a message preceded by the simulation time and the name of the arrival.
}

