% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{needs_attrs}
\alias{needs_attrs}
\title{Checks if attributes should be supplied}
\usage{
needs_attrs(variable)
}
\arguments{
\item{variable}{the variable to check}
}
\value{
a boolean
}
\description{
Checks if attributes should be supplied
}

