% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator-methods.R
\name{now}
\alias{now}
\title{Simulation Time}
\usage{
now(.env)
}
\arguments{
\item{.env}{the simulation environment.}
}
\value{
Returns a numeric value.
}
\description{
Get the current simulation time.
}
\seealso{
\code{\link{peek}}.
}
