% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{onestep}
\alias{onestep}
\title{Step the simulation}
\usage{
onestep(env)
}
\arguments{
\item{env}{the simulation environment.}
}
\value{
Returns the simulation environment.
}
\description{
Processes the next event.
}
\seealso{
\code{\link{reset}}, \code{\link{run}}.
}

