% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_NNLS_functions.R
\name{par_draw}
\alias{par_draw}
\title{Draw parameters}
\usage{
par_draw(updates, current_pars, kineticParPrior)
}
\arguments{
\item{updates}{the parameters corresponding to indeces of kineticParPrior}

\item{current_pars}{current kinetic parameters some which will be overwritten}

\item{kineticParPrior}{parameters for the proposal distribution of each kinetic parameter}
}
\value{
updated kinetic parameters
}
\description{
Update parameters using their prior (given by kineticParPrior) - update those those parameters
whose index is in "updates"
}
\details{
Options for these parameters are:
-@-@ unif: uniform in log-space: par_1 = lowerbound, par_2 = upperbound. draw in log2 space and exponentiate back to linear space
-@-@ norm: lognormal: in log2 space draw a value from mean = par_1, sd = par_2
-@-@ SpSl: spike and slab (In log2 space: the spike is a point mass at zero with p = par_3, the slab is a normal with mean = par_1 and sd = par_2)
}

