% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QP_FBA_functions.R
\name{parse_custom}
\alias{parse_custom}
\title{Parse custom metabolic model extensions}
\usage{
parse_custom(folder, file)
}
\arguments{
\item{customRx}{directory of customRxns.txt}
}
\value{
list containing reactions and metabolites that were added to the base metabolic model
}
\description{
Add additional reactions, metabolites, etc to metabolic model
}

