% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_attributes}
\alias{plot_attributes}
\title{Plot evolution of attribute data
 
Plot the evolution of user-supplied attribute data.}
\usage{
plot_attributes(envs, keys = c())
}
\arguments{
\item{envs}{a single simmer environment or a list of environments representing several replications}

\item{keys}{the keys of attributes you want to plot (if left empty, all attributes are shown)}
}
\description{
Plot evolution of attribute data
 
Plot the evolution of user-supplied attribute data.
}

