% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_resource_usage}
\alias{plot_resource_usage}
\title{plot utilization of a resource over time}
\usage{
plot_resource_usage(sim_obj, resource_name, replication = FALSE)
}
\arguments{
\item{sim_obj}{the simulation object}

\item{resource_name}{the name of the resource (character value)}

\item{replication}{specify to plot only a specific replication (default=FALSE)}
}
\description{
plot the usage of a resource over the simulation time frame
}

