% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_resource_utilization}
\alias{plot_resource_utilization}
\title{plot utilization of resources}
\usage{
plot_resource_utilization(sim_obj, resources)
}
\arguments{
\item{sim_obj}{the simulation object}

\item{resources}{a character vector with at least one resource specified - e.g. "c('res1','res2')"}
}
\description{
plot the utilization of specified resources in the simulation
}

