% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_NNLS_functions.R
\name{populate_reactionEqns}
\alias{populate_reactionEqns}
\title{Populate reaction equations}
\usage{
populate_reactionEqns()
}
\value{
generates the reactionForms.Rds file
}
\description{
Not meant for users. Setup a subset of reaction forms for package including rMM reaction forms,
significant complex regulation (2 regulators or cooperativity) and either significant single regulators
(or best regulators for complex regulation)
}

