% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity-methods.R
\name{print.externalptr}
\alias{print.externalptr}
\alias{print_activity}
\title{Print an activity}
\usage{
\method{print}{externalptr}(x, ...)

print_activity(activity)
}
\arguments{
\item{x}{an external pointer to the activity.}

\item{...}{arguments to be passed to or from other methods.}

\item{activity}{an external pointer to the activity.}
}
\description{
It can be used to visualise an activity's internal structure.
}
\seealso{
\code{\link{head.trajectory}}, \code{\link{tail.trajectory}},
\code{\link{get_next_activity}}, \code{\link{get_prev_activity}}.
}

