% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{print_activity}
\alias{print_activity}
\title{Print an activity}
\usage{
print_activity(activity)
}
\arguments{
\item{activity}{an external pointer to the activity.}
}
\description{
It can be used to visualise an activity's internal structure.
}
\seealso{
\code{\link{get_head}}, \code{\link{get_tail}}, 
\code{\link{get_next_activity}}, \code{\link{get_prev_activity}}.
}

