% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactionForms}
\alias{reactionForms}
\title{Data needed to evaluate each reaction form}
\format{a list with 93 reaction equations}
\usage{
reactionForms
}
\description{
Data needed to evaluate each reaction form
}
\keyword{datasets}

