% Generated by roxygen2 (4.0.1): do not edit by hand
\name{release_resource}
\alias{release_resource}
\title{release n amount of resource x}
\usage{
release_resource(trajectory_obj, resource_name, amount,
  description = "undefined", successor_id = NA, event_id = NA)
}
\arguments{
\item{trajectory_obj}{the trajectory object}

\item{resource_name}{the name of the resource}

\item{amount}{the amount of \emph{resource_name} to be seized}
}
\description{
release n amount of resource x
}

