% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{release_selected}
\alias{release_selected}
\title{Add a release activity}
\usage{
release_selected(traj, amount = 1, id = 0)
}
\arguments{
\item{traj}{the trajectory object.}

\item{amount}{the amount to release, accepts either a callable object (a function) or a numeric value.}

\item{id}{selection identifier for nested usage.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of releasing a previously selected resource to the tail of a trajectory.
}
\seealso{
\code{\link{seize}}, \code{\link{release}}, \code{\link{seize_selected}}, 
\code{\link{select}}.
}

