% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event.R
\docType{data}
\name{ReleaseEvent}
\alias{ReleaseEvent}
\title{ReleaseEvent}
\format{An \code{\link{R6Class}} generator object}
\usage{
ReleaseEvent
}
\description{
Release a resource by an amount
}
\section{Fields}{

\describe{
\item{\code{resource}}{the resource name}

\item{\code{amount}}{the amount to release}
}}
\keyword{datasets}

