% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replicator}
\alias{replicator}
\title{Replicate the simulation n times}
\usage{
replicator(sim_obj, n_replications)
}
\arguments{
\item{sim_obj}{the Simulator object}

\item{n_replications}{the number of desired replications}
}
\description{
Replicate the simulation n times
}

