% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{reset}
\alias{reset}
\title{Reset a simulator}
\usage{
reset(env)
}
\arguments{
\item{env}{the simulation environment.}
}
\value{
Returns the simulation environment.
}
\description{
Resets the following components of a simulation environment: 
time, event queue, resources, generators and statistics.
}
\seealso{
Other methods to deal with a simulation environment:
\link{simmer}, \link{now}, \link{peek}, \link{step}, \link{run}, 
\link{add_resource}, \link{add_generator}, \link{get_mon_arrivals}, \link{get_mon_resources}, 
\link{get_capacity}, \link{get_queue_size}, \link{get_server_count}, \link{get_queue_count}.
}

