% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rMech_summary_table}
\alias{rMech_summary_table}
\title{Summary of reaction forms}
\format{a data frame with 93 rows and 3 variables
\describe{
\item{reaction}{Systematic ID of reaction}
\item{rMech}{ID of reaction form}
\itme{modelType}{Type of regulation}
}}
\usage{
rMech_summary_table
}
\description{
Summary of reaction forms
}
\keyword{datasets}

