% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{run}
\alias{run}
\title{Run the simulation}
\usage{
run(env, until = 1000)
}
\arguments{
\item{env}{the simulation environment.}

\item{until}{stop time.}
}
\value{
Returns the simulation environment.
}
\description{
Executes steps until the given criterion.
}
\seealso{
\code{\link{reset}}, \code{\link{onestep}}.
}

