% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_funcs.R
\name{schedule}
\alias{schedule}
\title{Generate a scheduling object}
\usage{
schedule(timetable, period, values)
}
\arguments{
\item{timetable}{absolute points in time in which the desired value changes.}

\item{period}{period of repetition.}

\item{values}{one value for each point in time.}
}
\value{
Returns a Schedule object.
}
\description{
Resource convenience function to generate a scheduling object from a timetable specification.
}
\examples{
# Schedule 3 units from 8 to 16 h
#          2 units from 16 to 24 h
#          1 units from 24 to 8 h
my_schedule <- schedule(c(8, 16, 24), 24, c(3, 2, 1))
}

