% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{seize}
\alias{seize}
\title{Add a seize activity}
\usage{
seize(traj, resource, amount = 1, ...)
}
\arguments{
\item{traj}{the trajectory object.}

\item{resource}{the name of the resource.}

\item{amount}{the amount to seize, accepts either a callable object (a function) or a numeric value.}

\item{...}{unused arguments}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of seizing a resource to the tail of a trajectory.
}
\seealso{
\code{\link{release}}, \code{\link{seize_selected}}, 
\code{\link{release_selected}}, \code{\link{select}}.
}

