% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event.R
\docType{data}
\name{SeizeEvent}
\alias{SeizeEvent}
\title{SeizeEvent}
\format{An \code{\link{R6Class}} generator object}
\usage{
SeizeEvent
}
\description{
Seize a resource by an amount
}
\section{Fields}{

\describe{
\item{\code{resource}}{the resource name}

\item{\code{amount}}{the amount to seize}
}}
\keyword{datasets}

