% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{set_queue_size}
\alias{set_queue_size}
\title{Set the queue size}
\usage{
set_queue_size(env, name, value)
}
\arguments{
\item{env}{the simulation environment.}

\item{name}{the name of the resource.}

\item{value}{new queue size.}
}
\value{
Returns the simulation environment.
}
\description{
Sets a resource's queue size by name.
}
\seealso{
\code{\link{get_capacity}}, \code{\link{get_queue_size}}, \code{\link{set_capacity}},
\code{\link{get_server_count}}, \code{\link{get_queue_count}}.
}

