% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{set_trajectory}
\alias{set_trajectory}
\alias{set_distribution}
\title{Set Generator Parameters}
\usage{
set_trajectory(.trj, generator, trajectory)

set_distribution(.trj, generator, distribution)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{generator}{the name of the generator or a function returning a name.}

\item{trajectory}{the trajectory that the generated arrivals will follow.}

\item{distribution}{a function modelling the interarrival times (returning a
negative value stops the generator).}
}
\value{
Returns the trajectory object.
}
\description{
Activities for modifying a generator's trajectory or distribution by name.
}
\seealso{
\code{\link{activate}}, \code{\link{deactivate}}.
}
