% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QP_FBA_functions.R
\name{setup_FBA_constraints}
\alias{setup_FBA_constraints}
\title{Setup FBA constraints}
\usage{
setup_FBA_constraints(return_infeasible = F, infRxMet = NULL)
}
\arguments{
\item{return_infeasible}{logical: if TRUE then only return metabolite and reaction IDs to be removed.
If FALSE then determine all equality and inequality constraints.}

\item{infRxMet}{a list of metabolites and reactions to remove}
}
\value{
write optimization constraints to global environment
}
\description{
either determine reactions that cannot carry flux given stoichiometry, boundary fluxes and directionality or write all
equality and inequality constraints to the global environment.
}
\details{
If return_infeasible is TRUE then a data.frame of reactions that cannot carry flux (and associated metabolites)
will be returned providing this data.frame as infRxMet will exclude these reactions and species from subsequent steps.
}

