% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{show_activity}
\alias{show_activity}
\title{Show an activity}
\usage{
show_activity(activity)
}
\arguments{
\item{activity}{an external pointer to the activity.}
}
\description{
It can be used to visualise an activity's internal structure.
}
\seealso{
Other methods to deal with activities: 
\link{get_next_activity}, \link{get_prev_activity}.
}

