% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trajectory.R
\name{show_trajectory}
\alias{show_trajectory}
\title{Show a trajectory}
\usage{
show_trajectory(traj)
}
\arguments{
\item{traj}{the trajectory object.}
}
\description{
It can be used to visualise a trajectory's internal structure.
}
\seealso{
Other methods to deal with trajectories:
\link{create_trajectory}, \link{get_head}, \link{get_tail},
\link{get_n_activities}, \link{seize}, \link{release}, \link{timeout},
\link{branch}, \link{rollback}.
}

