% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulator.R
\docType{data}
\name{Simmer.Env}
\alias{Simmer.Env}
\title{Simmer.Env}
\format{An \code{\link{R6Class}} generator object}
\usage{
Simmer.Env
}
\description{
Simulation environment
}
\section{Fields}{

\describe{
\item{\code{name}}{environment name}
}}
\keyword{datasets}

