% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{Simmer.make_closure}
\alias{Simmer.make_closure}
\title{Simmer.make_closure}
\usage{
Simmer.make_closure(is_gen, expr)
}
\arguments{
\item{is_gen}{must be TRUE if the process is a generator (has a while loop), FALSE otherwise}

\item{expr}{the function to convert}
}
\value{
a process implemented as an R closure
}
\description{
convert a SimPy-like process into an R closure
}

