% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulator.R
\docType{data}
\name{Simmer.wrap}
\alias{Simmer.wrap}
\title{Simmer.wrap}
\format{An \code{\link{R6Class}} generator object}
\usage{
Simmer.wrap
}
\description{
Extracts the simulation data from a Simmer object making it accessible through the same methods
}
\section{Fields}{

\describe{
\item{\code{simmer}}{a simmer object}
}}
\keyword{datasets}

