% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{timeout}
\alias{timeout}
\title{Add a timeout activity}
\usage{
timeout(traj, task, provide_attrs = FALSE)
}
\arguments{
\item{traj}{the trajectory object.}

\item{task}{a callable object (a function) that returns a numeric value 
(negative values are automatically converted to positive ones)}

\item{provide_attrs}{should the attributes be provided to the value function (value has to be a function which accept a parameter)}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of executing any user-defined task and setting
an associated delay to the tail of a trajectory.
}
\seealso{
Other methods to deal with trajectories:
\link{create_trajectory}, \link{show_trajectory}, \link{get_head},
\link{get_tail}, \link{get_n_activities}, \link{seize}, \link{release}, 
\link{branch}, \link{rollback}.
}

