% Generated by roxygen2 (4.0.1): do not edit by hand
\name{timeout_entity}
\alias{timeout_entity}
\title{timeout entity for x duration}
\usage{
timeout_entity(trajectory_obj, duration, ...)
}
\arguments{
\item{trajectory_obj}{the trajectory object}

\item{duration}{timeout x time units}
}
\description{
timeout entity for x duration
}

