% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event.R
\docType{data}
\name{TimeoutEvent}
\alias{TimeoutEvent}
\title{TimeoutEvent}
\format{An \code{\link{R6Class}} generator object}
\usage{
TimeoutEvent
}
\description{
Wait for a given delay
}
\section{Fields}{

\describe{
\item{\code{duration}}{a function that generates a delay}
}}
\keyword{datasets}

