% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{TimeValueMonitor-class}
\alias{TimeValueMonitor}
\alias{TimeValueMonitor-class}
\title{TimeValueMonitor class.}
\description{
Holds data related to specific point in simulation time. E.g. holds values for resource usage at time x.
}

