% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trajectory.R
\docType{data}
\name{Trajectory}
\alias{Trajectory}
\title{Trajectory}
\format{An \code{\link{R6Class}} generator object}
\usage{
Trajectory
}
\description{
Trajectory object
}
\section{Fields}{

\describe{
\item{\code{name}}{trajectory name}
}}
\examples{
t0 <- Trajectory$new("my trajectory") $
  seize("server", 1) $
  timeout(function() rexp(1, 1)) $
  release("server", 1)
}
\keyword{datasets}

