% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{Wrap a simulation environment}
\usage{
wrap(env)
}
\arguments{
\item{env}{the simulation environment.}
}
\value{
Returns a simulation wrapper.
}
\description{
This function extracts the monitored data from a simulation environment 
making it accessible through the same methods. Only useful if you want 
to parallelize heavy replicas (see the example below), because the C++ 
simulation backend is destroyed when the threads exit.
}
\examples{
library(parallel)

mm1 <- create_trajectory() \%>\%
  seize("server", 1) \%>\%
  timeout(function() rexp(1, 2)) \%>\%
  release("server", 1)

envs <- mclapply(1:4, function(i) {
  simmer("M/M/1 example") \%>\%
    add_resource("server", 1) \%>\%
    add_generator("customer", mm1, function() rexp(1, 1)) \%>\%
    run(100) \%>\%
    wrap()
})

plot_resource_usage(envs, "server")
}
\seealso{
Other methods for dealing with a simulation wrapper:
\link{get_mon_arrivals}, \link{get_mon_attributes}, \link{get_mon_resources}, 
\link{get_n_generated}, \link{get_capacity}, \link{get_queue_size},
\link{get_server_count}, \link{get_queue_count}.
}

