% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{activate}
\alias{activate}
\alias{deactivate}
\title{Add a activate/deactivate activity}
\usage{
activate(.trj, generator)

deactivate(.trj, generator)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{generator}{the name of the generator or a function returning a name.}
}
\value{
Returns the trajectory object.
}
\description{
Activate or deactivate the generation of arrivals by name.
}
\seealso{
\code{\link{set_trajectory}}, \code{\link{set_distribution}}.
}

