% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_entity}
\alias{add_entity}
\title{Add an individual entity}
\usage{
add_entity(sim_obj, name, trajectory_name, early_start = 0)
}
\arguments{
\item{sim_obj}{the Simulator object}

\item{name}{the name of the entity}

\item{trajectory_name}{the name of the trajectory to use}

\item{early_start}{the time at which the entity is activated}
}
\description{
Add an individual entity
}

