% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{add_generator}
\alias{add_generator}
\title{Add a generator}
\usage{
add_generator(env, name_prefix, trajectory, dist, mon = TRUE)
}
\arguments{
\item{env}{the simulation environment.}

\item{name_prefix}{the name prefix of the generated arrivals.}

\item{trajectory}{the trajectory that the generated arrivals will follow (see \link{create_trajectory}).}

\item{dist}{a function modelling the interarrival times (returning a negative value stops the generator).}

\item{mon}{whether the simulator must monitor the generated arrivals or not.}
}
\value{
Returns the simulation environment.
}
\description{
Adds a generator to a simulation environment.
}
\seealso{
Other methods to deal with a simulation environment:
\link{simmer}, \link{now}, \link{peek}, \link{peek}, \link{onestep}, 
\link{run}, \link{add_resource}, \link{get_mon_arrivals}, \link{get_mon_resources}, 
\link{get_capacity}, \link{get_queue_size}, \link{get_server_count}, \link{get_queue_count}.
}

