% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_resource}
\alias{add_resource}
\title{Add a resource}
\usage{
add_resource(sim_obj, name, capacity = 1)
}
\arguments{
\item{sim_obj}{the Simulator object}

\item{name}{specify the name of the resource}

\item{the}{capacity of the resource}
}
\description{
Add a resource
}

