% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_resource}
\alias{add_resource}
\title{Add a resource to the simulation object}
\usage{
add_resource(sim_obj, name, capacity)
}
\arguments{
\item{sim_obj}{the simulation object}

\item{name}{the name of the resource}

\item{capacity}{the capacity of the resource}
}
\description{
Add a resource to the simulation object
}

