% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulator.R
\name{add_resource}
\alias{add_resource}
\title{Add a resource to the simulation object}
\usage{
add_resource(sim_obj, name, capacity, queue_size = Inf)
}
\arguments{
\item{sim_obj}{the simulation object}

\item{name}{the name of the resource}

\item{capacity}{the capacity of the resource}

\item{queue_size}{the queue size of the resource}
}
\description{
Add a resource to the simulation object
}

