% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{add_resource}
\alias{add_resource}
\title{Add a resource}
\usage{
add_resource(env, name, capacity = 1, queue_size = Inf, mon = TRUE)
}
\arguments{
\item{env}{the simulation environment.}

\item{name}{the name of the resource.}

\item{capacity}{the capacity of the server.}

\item{queue_size}{the size of the queue.}

\item{mon}{whether the simulator must monitor this resource or not.}
}
\value{
Returns the simulation environment.
}
\description{
Adds a resource to a simulation environment.
}
\seealso{
Other methods to deal with a simulation environment:
\link{simmer}, \link{reset}, \link{now}, \link{peek}, \link{onestep}, \link{run}, 
\link{add_generator}, \link{get_mon_arrivals}, \link{get_mon_attributes},
\link{get_mon_resources}, \link{get_n_generated}, \link{get_capacity}, \link{get_queue_size},
\link{get_server_count}, \link{get_queue_count}.
}

