% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_seize_event}
\alias{add_seize_event}
\title{add a seize event to the trajectory object}
\usage{
add_seize_event(trajectory_obj, resource, amount)
}
\arguments{
\item{trajectory_obj}{the trajectory object}

\item{resource}{the resource as a character string}

\item{amount}{the amount of resources to seize}
}
\description{
add a seize event to the trajectory object
}

