% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_funcs.R
\name{at}
\alias{at}
\title{Arrivals at specific times}
\usage{
at(...)
}
\arguments{
\item{...}{a vector or multiple parameters of times at which to initiate an arrival.}
}
\value{
Returns a generator function.
}
\description{
Generator convenience function to generate arrivals at specific times.
}
\examples{
t0 <- create_trajectory() \%>\% timeout(0)
env <- simmer(verbose=T) \%>\%
  add_generator("dummy", t0, at(0, c(1,10,30), 40, 43)) \%>\%
  run(100)
}
\seealso{
\link{add_generator}
}

