% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{batch}
\alias{batch}
\alias{split}
\title{Add a batch/split activity}
\usage{
batch(traj, n, timeout = NULL, rule = NULL, permanent = FALSE)

split(traj)
}
\arguments{
\item{traj}{the trajectory object.}

\item{n}{batch size, accepts a numeric.}

\item{timeout}{optional time limit after which the batch will continue
even if the batch size has not been fulfilled, accepts a numeric.}

\item{rule}{an optional callable object (a function) which will be applied to 
every arrival to determine whether it should be included into the batch, thus}

\item{permanent}{if \code{TRUE}, batches cannot be split.}
}
\value{
The trajectory object.
}
\description{
Collect a number of arrivals before they can continue processing
or split a previously established batch.
}

