% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trajectory.R
\docType{data}
\name{Branch}
\alias{Branch}
\title{Branch}
\format{An \code{\link{R6Class}} generator object}
\usage{
Branch
}
\description{
Break a trajectory in several branches
}
\keyword{datasets}

