% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{branch}
\alias{branch}
\title{Add a branch activity}
\usage{
branch(traj, option, merge, ...)
}
\arguments{
\item{traj}{the trajectory object.}

\item{option}{a callable object (a function) that must return an integer 
between 1 and n; it will be used by the arrivals to select a path to follow.}

\item{merge}{a vector of n booleans that indicate whether the arrival must continue 
executing activities after each path or not.}

\item{...}{n trajectory objects describing each path.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity that defines n alternative paths to the tail of a trajectory.
}
\seealso{
Other methods to deal with trajectories:
\link{create_trajectory}, \link{show_trajectory}, \link{get_head},
\link{get_tail}, \link{get_n_activities}, \link{seize}, \link{release}, \link{timeout}.
}

