% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{branch}
\alias{branch}
\title{Add a branch activity}
\usage{
branch(traj, option, continue, ..., merge = "_deprecated")
}
\arguments{
\item{traj}{the trajectory object.}

\item{option}{a callable object (a function) that must return an integer 
between 1 and \code{n}; it will be used by the arrivals to select a path to follow.}

\item{continue}{a vector of \code{n} booleans that indicate whether the arrival must 
continue executing activities after each path or not.}

\item{...}{\code{n} trajectory objects describing each path.}

\item{merge}{(deprecated) same as \code{continue}, for compatibility reasons. Support
for this argument will be dropped in upcoming versions.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity that defines n alternative paths to the tail of a trajectory.
}
\seealso{
Other methods for dealing with trajectories:
\link{create_trajectory}, \link{get_head}, \link{get_tail}, 
\link{get_n_activities}, \link{join}, \link{seize}, \link{release}, 
\link{set_attribute}, \link{timeout}, \link{rollback}.
}

