% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{branch}
\alias{branch}
\title{Add a branch activity}
\usage{
branch(traj, option, continue, ...)
}
\arguments{
\item{traj}{the trajectory object.}

\item{option}{a callable object (a function) which must return an integer between
\code{0} and \code{N}. A return value equal to \code{0} skips the branch and
continues to the next activity. A returning value between \code{1} to \code{N}
makes the arrival to follow the corresponding sub-trajectory.}

\item{continue}{a vector of \code{N} booleans that indicate whether the arrival must
continue to the main trajectory after each sub-trajectory or not.}

\item{...}{\code{N} trajectory objects describing each sub-trajectory.}
}
\value{
Returns the trajectory object.
}
\description{
Define a fork with \code{N} alternative sub-trajectories.
}

