% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_entity}
\alias{create_entity}
\title{Creates an entity object}
\usage{
create_entity(name = "anonymous", activation_time = 0)
}
\arguments{
\item{entity_name}{the name of the entity (defaults to 'anonymous')}
}
\description{
Creates an entity object
}

