% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entity.R
\name{create_entity}
\alias{create_entity}
\title{Creates an entity object}
\usage{
create_entity(name = "anonymous", activation_time = 0)
}
\arguments{
\item{name}{the name of the entity (defaults to 'anonymous')}

\item{activation_time}{the time at which the entity is activated}
}
\description{
Creates an entity object
}

