% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_simulator}
\alias{create_simulator}
\title{#Creates a simulator object
#
#}
\usage{
create_simulator(sim_name = "anonymous", n = 1, until = Inf,
  verbose = FALSE)
}
\arguments{
\item{sim_name}{the name of the simulator (defaults to 'anonymous')
#}

\item{n}{the number of replications}

\item{until}{the maximum run time of the simulation
#}

\item{verbose}{show log messagses
#}

\item{name}{the name of the simulator (defaults to 'anonymous')}

\item{until}{the maximum run time of the simulation}

\item{verbose}{show log messages}
}
\description{
#Creates a simulator object
#
#
}

