% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trajectory.R
\name{create_trajectory}
\alias{create_trajectory}
\title{create a trajectory object}
\usage{
create_trajectory(name)
}
\arguments{
\item{name}{the name of the trajectory}
}
\description{
create a trajectory object
}

