% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_funcs.R
\name{every}
\alias{every}
\title{Arrivals every specific interval}
\usage{
every(...)
}
\arguments{
\item{...}{a vector or multiple parameters of intervals between arrivals.}
}
\value{
Returns a generator function.
}
\description{
Generator convenience function to generate arrivals every specific interval.
When the generator reaches the last interval, it starts again from the first one.
}
\examples{
t0 <- create_trajectory() \%>\% timeout(0)
env <- simmer(verbose=T) \%>\%
  add_generator("dummy", t0, every(1, 2, 1)) \%>\%
  run(10)
}
\seealso{
\link{add_generator}, \link{at}, \link{from}, 
\link{to}, \link{from_to}.
}

