% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{from}
\alias{from}
\title{Generate arrivals starting at a specified time}
\usage{
from(start_time, dist, arrive = TRUE)
}
\arguments{
\item{start_time}{the time at which to launch the initial arrival.}

\item{dist}{a function modelling the interarrival times.}

\item{arrive}{if set to \code{TRUE} (default) the first arrival will be
generated at \code{start_time} and will follow \code{dist} from then on.
If set to \code{FALSE}, will initiate \code{dist} at \code{start_time}
(and the first arrival will most likely start at a time later than
\code{start_time}).}
}
\value{
Returns a generator function.
}
\description{
Generator convenience function to generate inter-arrivals with a specified start time.
}
\examples{
t0 <- trajectory() \%>\%
  timeout(0)

simmer() \%>\%
  add_generator("dummy", t0, from(5, function() runif(1, 1, 2))) \%>\%
  run(10) \%>\%
  get_mon_arrivals()
}
\seealso{
\code{\link{add_generator}}.
}
