% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator-methods.R
\name{get_capacity}
\alias{get_capacity}
\alias{get_queue_size}
\alias{get_server_count}
\alias{get_queue_count}
\title{Get resource parameters}
\usage{
get_capacity(.env, resource)

get_queue_size(.env, resource)

get_server_count(.env, resource)

get_queue_count(.env, resource)
}
\arguments{
\item{.env}{the simulation environment.}

\item{resource}{the name of the resource.}
}
\value{
Return a numeric value.
}
\description{
Getters for resources: server capacity/count and queue size/count.
}
\seealso{
\code{\link{set_capacity}}, \code{\link{set_queue_size}}.
}
