% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_capacity}
\alias{get_capacity}
\title{Get the capacity}
\usage{
get_capacity(env, name)
}
\arguments{
\item{env}{the simulation environment.}

\item{name}{the name of the resource.}
}
\value{
Returns a numeric value.
}
\description{
Gets the capacity of a resource by name.
}
\seealso{
Other methods for dealing with a simulation environment:
\link{simmer}, \link{reset}, \link{now}, \link{peek}, \link{onestep}, \link{run}, 
\link{add_resource}, \link{add_generator}, \link{get_mon_arrivals}, \link{get_mon_attributes},
\link{get_mon_resources}, \link{get_n_generated}, \link{get_queue_size},
\link{get_server_count}, \link{get_queue_count}.
}

