% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer-methods.R
\name{get_capacity}
\alias{get_capacity}
\alias{get_capacity_selected}
\alias{get_queue_size}
\alias{get_queue_size_selected}
\alias{get_server_count}
\alias{get_server_count_selected}
\alias{get_queue_count}
\alias{get_queue_count_selected}
\alias{get_seized}
\alias{get_seized_selected}
\alias{get_activity_time}
\alias{get_activity_time_selected}
\alias{get_selected}
\title{Get Resource Parameters}
\usage{
get_capacity(.env, resources)

get_capacity_selected(.env, id = 0)

get_queue_size(.env, resources)

get_queue_size_selected(.env, id = 0)

get_server_count(.env, resources)

get_server_count_selected(.env, id = 0)

get_queue_count(.env, resources)

get_queue_count_selected(.env, id = 0)

get_seized(.env, resources)

get_seized_selected(.env, id = 0)

get_activity_time(.env, resources)

get_activity_time_selected(.env, id = 0)

get_selected(.env, id = 0)
}
\arguments{
\item{.env}{the simulation environment.}

\item{resources}{one or more resource names.}

\item{id}{selection identifier for nested usage.}
}
\value{
Return a vector (character for \code{get_selected}, numeric for the
rest of them).
}
\description{
Getters for resources: server capacity/count and queue size/count, seized
amount, activity time, and selected resources.
}
\details{
If no resources are provided to \code{get_activity_time}, the
overall activity time is reported.
}
\seealso{
\code{\link{get_resources}}, \code{\link{set_capacity}},
\code{\link{set_queue_size}}, \code{\link{seize}}, \code{\link{timeout}}.
}
