% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_mon}
\alias{get_mon}
\alias{get_mon_arrivals}
\alias{get_mon_attributes}
\alias{get_mon_resources}
\title{Get statistics}
\usage{
get_mon_arrivals(envs, per_resource = FALSE)

get_mon_attributes(envs)

get_mon_resources(envs, data = c("counts", "limits"))
}
\arguments{
\item{envs}{the simulation environment (or a list of environments).}

\item{per_resource}{whether the activity should be reported on a per-resource basis (by default: FALSE).}

\item{data}{whether to retrieve the "counts", the "limits" or both.}
}
\value{
Return a data frame.
}
\description{
Simulator getters for obtaining monitored data (if any) about arrivals, attributes and resources.
}

