% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_mon_arrivals}
\alias{get_mon_arrivals}
\title{Get arrival statistics}
\usage{
get_mon_arrivals(envs, per_resource = FALSE)
}
\arguments{
\item{envs}{the simulation environment (or a list of environments).}

\item{per_resource}{whether the activity should be reported on a per-resource basis (by default: FALSE).}
}
\value{
Returns a data frame.
}
\description{
Gets the arrivals' monitored data (if any).
}
\seealso{
\code{\link{get_mon_attributes}}, \code{\link{get_mon_resources}}.
}

