% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_mon_attributes}
\alias{get_mon_attributes}
\title{Get attribute statistics}
\usage{
get_mon_attributes(env)
}
\arguments{
\item{env}{the simulation environment (or a list of environments).}
}
\value{
Returns a data frame.
}
\description{
Gets the arrivals' attributes over time (if any).
}
\seealso{
Other methods for dealing with a simulation environment:
\link{simmer}, \link{reset}, \link{now}, \link{peek}, \link{onestep}, \link{run}, 
\link{add_resource}, \link{add_generator}, \link{get_mon_arrivals},
\link{get_mon_resources}, \link{get_n_generated}, \link{get_capacity}, \link{get_queue_size},
\link{get_server_count}, \link{get_queue_count}.
}

