% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{get_n_activities}
\alias{get_n_activities}
\title{Get the number of activities}
\usage{
get_n_activities(trajectory)
}
\arguments{
\item{trajectory}{the trajectory object.}
}
\value{
Returns the number of activities in the trajectory.
}
\description{
Trajectory getter for obtaining the total number of activities defined inside it.
}
\seealso{
\code{\link{get_head}}, \code{\link{get_tail}}, \code{\link{join}}.
}

