% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_n_generated}
\alias{get_n_generated}
\title{Get the number of arrivals generated}
\usage{
get_n_generated(env, name)
}
\arguments{
\item{env}{the simulation environment.}

\item{name}{the name of the generator.}
}
\value{
Returns a numeric value.
}
\description{
Gets the number of arrivals generated by a generator by name.
}
\seealso{
Other methods to deal with a simulation environment:
\link{simmer}, \link{reset}, \link{now}, \link{peek}, \link{onestep}, \link{run}, 
\link{add_resource}, \link{add_generator}, \link{get_mon_arrivals}, \link{get_mon_attributes},
\link{get_mon_resources}, \link{get_capacity}, \link{get_queue_size},
\link{get_server_count}, \link{get_queue_count}.
}

