% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{get_n_generated}
\alias{get_n_generated}
\title{Get the number of arrivals generated}
\usage{
get_n_generated(env, name)
}
\arguments{
\item{env}{the simulation environment.}

\item{name}{the name of the generator.}
}
\value{
Returns a numeric value.
}
\description{
Simulator getter for obtaining the number of arrivals generated by a generator by name.
}

