% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer-methods.R
\name{get_n_generated}
\alias{get_n_generated}
\alias{get_name}
\alias{get_attribute}
\alias{get_global}
\alias{get_prioritization}
\title{Get Process Parameters}
\usage{
get_n_generated(.env, source)

get_name(.env)

get_attribute(.env, keys, global = FALSE)

get_global(.env, keys)

get_prioritization(.env)
}
\arguments{
\item{.env}{the simulation environment.}

\item{source}{the name of the source.}

\item{keys}{the attribute name(s).}

\item{global}{\code{global=TRUE} is deprecated. Use \code{*_global} instead.}
}
\description{
Getters for processes (sources and arrivals) number of arrivals generated
by a source, the name of the active arrival, an attribute from the active
arrival or a global one, and prioritization values.
}
\details{
\code{get_n_generated} returns the number of arrivals generated by a
given source.

\code{get_name} returns the number of the running arrival. \code{get_attribute}
returns a running arrival's attributes. If a provided key was not previously
set, it returns a missing value. \code{get_global} returns a global attribute.
\code{get_prioritization} returns a running arrival's prioritization values.
\code{get_name}, \code{get_attribute} and \code{get_prioritization} are meant
to be used inside a trajectory; otherwise, there will be no arrival running
and these functions will throw an error.
}
\seealso{
\code{\link{set_attribute}}, \code{\link{set_global}},
\code{\link{set_prioritization}}.
}
