% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_prev_activity}
\alias{get_prev_activity}
\title{Get the previous activity}
\usage{
get_prev_activity(activity)
}
\arguments{
\item{activity}{an external pointer to the activity.}
}
\value{
An external pointer to an activity object.
}
\description{
It takes an external pointer to an activity an returns the previous activity.
}
\seealso{
\code{\link{print_activity}}, \code{\link{get_next_activity}}.
}

