% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{get_tail}
\alias{get_tail}
\title{Get the last activity}
\usage{
get_tail(traj)
}
\arguments{
\item{traj}{the trajectory object.}
}
\value{
An external pointer to an activity object.
}
\description{
It can be used to get the pointer to a trajectory's last activity.
}
\seealso{
Other methods to deal with trajectories:
\link{create_trajectory}, \link{get_head},
\link{get_n_activities}, \link{seize}, \link{release}, \link{timeout}, 
\link{set_attribute}, \link{branch}, \link{rollback}.
}

