% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{head.trajectory}
\alias{head.trajectory}
\alias{tail.trajectory}
\title{Return the first/last activity}
\usage{
\method{head}{trajectory}(x, n = 1L, ...)

\method{tail}{trajectory}(x, n = 1L, ...)
}
\arguments{
\item{x}{the trajectory object.}

\item{n}{a single integer. If positive, size for the resulting
    object: number of elements for a vector (including lists), rows for
    a matrix or data frame or lines for a function. If negative, all but
    the \code{n} last/first number of elements of \code{x}.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
Returns a new trajectory object.
}
\description{
Returns the first or last parts of a trajectory object.
}
\seealso{
\code{\link{length.trajectory}}, \code{\link{get_n_activities}},
\code{\link{join}}.
}

