% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{join}
\alias{join}
\title{Join trajectories}
\usage{
join(...)
}
\arguments{
\item{...}{trajectory objects.}
}
\value{
A new trajectory object.
}
\description{
Concatenate any number of trajectories in the order specified.
}
\examples{
t1 <- create_trajectory() \%>\% seize("dummy", 1)
t2 <- create_trajectory() \%>\% timeout(1)
t3 <- create_trajectory() \%>\% release("dummy", 1)

join(t1, t2, t3)

create_trajectory() \%>\%
  join(t1) \%>\%
  timeout(1) \%>\%
  join(t3)
}
\seealso{
\code{\link{get_head}}, \code{\link{get_tail}}, \code{\link{get_n_activities}}.
}

