% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{leave}
\alias{leave}
\title{Add a leave activity}
\usage{
leave(traj, prob)
}
\arguments{
\item{traj}{the trajectory object.}

\item{prob}{a probability or a function returning a probability.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity that terminates the trajectory with some probability.
}
\seealso{
Other methods for dealing with trajectories:
\link{create_trajectory}, \link{get_head}, \link{get_tail}, 
\link{get_n_activities}, \link{join}, \link{seize}, \link{release}, 
\link{set_attribute}, \link{timeout}, \link{branch}, \link{rollback},
\link{seize_selected}, \link{release_selected}, \link{select}.
}

