% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{leave}
\alias{leave}
\title{Add a leave activity}
\usage{
leave(traj, prob)
}
\arguments{
\item{traj}{the trajectory object.}

\item{prob}{a probability or a function returning a probability.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity that terminates the trajectory with some probability.
}

