% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{length.trajectory}
\alias{get_n_activities}
\alias{length.trajectory}
\title{Number of activities in a trajectory}
\usage{
\method{length}{trajectory}(x)

get_n_activities(x)
}
\arguments{
\item{x}{the trajectory object.}
}
\value{
Returns a non-negative integer of length 1.
}
\description{
Get the number of activities in a trajectory. \code{length} returns the number
of first-level activities (sub-trajectories not included). \code{get_n_activities}
returns the total number of activities (sub-trajectories included).
}
\seealso{
\code{\link{head.trajectory}}, \code{\link{tail.trajectory}}, \code{\link{join}}.
}

