% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{log_}
\alias{log_}
\alias{stop_if}
\title{Debugging}
\usage{
log_(.trj, message, level = 0)

stop_if(.trj, condition)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{message}{the message to display, accepts either a string or a callable object
(a function) which must return a string.}

\item{level}{debugging level. The \code{message} will be printed if, and only if,
the \code{level} provided is less or equal to the \code{log_level} defined in the
simulation environment (see \code{\link{simmer}}).}
}
\value{
Returns the trajectory object.
}
\description{
Activities for displaying messages preceded by the simulation time and the
name of the arrival, and for stopping the simulation given some condition.
}
