% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\alias{monitor_mem}
\alias{monitor_csv}
\title{Create a Monitor}
\usage{
monitor_mem()

monitor_csv(path = tempdir(), reader = read.csv, sep = ",", ...)
}
\description{
Create a Monitor
}
