% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{now}
\alias{now}
\title{Get the current time}
\usage{
now(env)
}
\arguments{
\item{env}{the simulation environment.}
}
\value{
Returns a numeric value.
}
\description{
Gets the current simulation time.
}
\seealso{
Other methods to deal with a simulation environment:
\link{simmer}, \link{reset}, \link{peek}, \link{onestep}, \link{run}, 
\link{add_resource}, \link{add_generator}, \link{get_mon_arrivals}, \link{get_mon_resources}, 
\link{get_capacity}, \link{get_queue_size}, \link{get_server_count}, \link{get_queue_count}.
}

