% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{onestep}
\alias{onestep}
\title{Step the simulation}
\usage{
onestep(env)
}
\arguments{
\item{env}{the simulation environment.}
}
\value{
Returns the simulation environment.
}
\description{
Processes the next event.
}
\seealso{
Other methods to deal with a simulation environment:
\link{simmer}, \link{reset}, \link{now}, \link{peek}, \link{run}, 
\link{add_resource}, \link{add_generator}, \link{get_mon_arrivals}, \link{get_mon_resources}, 
\link{get_capacity}, \link{get_queue_size}, \link{get_server_count}, \link{get_queue_count}.
}

