% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator-methods.R
\name{peek}
\alias{peek}
\title{Peek next events}
\usage{
peek(.env, steps = 1, verbose = FALSE)
}
\arguments{
\item{.env}{the simulation environment.}

\item{steps}{number of steps to peek.}

\item{verbose}{show additional information (i.e., the name of the process) about future events.}
}
\value{
Returns numeric values if \code{verbose=F} and a data frame otherwise.
}
\description{
Look for future events in the event queue and (optionally) obtain info about them.
}
\seealso{
\code{\link{now}}.
}

