% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{peek}
\alias{peek}
\title{Peek next events}
\usage{
peek(env, steps = 1, verbose = FALSE)
}
\arguments{
\item{env}{the simulation environment.}

\item{steps}{number of steps to peek.}

\item{verbose}{show additional information (i.e., the name of the process) about future events.}
}
\value{
Returns numeric values if `verbose=F` and a data frame otherwise.
}
\description{
Looks for future events in the event queue.
}
\seealso{
\code{\link{now}}.
}

