% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_evolution_arrival_times}
\alias{plot_evolution_arrival_times}
\title{plot evolution of arrival times}
\usage{
plot_evolution_arrival_times(simmer, type = c("flow_time", "activity_time",
  "waiting_time"))
}
\arguments{
\item{simmer}{the simulation environment}

\item{type}{one of c("flow_time","activity_time","waiting_time")}
}
\description{
plot the evolution of arrival related times (flow, activity and waiting time)
}

