% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_evolution_arrival_times}
\alias{plot_evolution_arrival_times}
\title{Plot evolution of arrival times}
\usage{
plot_evolution_arrival_times(envs, type = c("activity_time", "waiting_time",
  "flow_time"))
}
\arguments{
\item{envs}{a single simmer environment or a list of environments representing several replications.}

\item{type}{one of \code{c("activity_time", "waiting_time", "flow_time")}.}
}
\value{
Returns a ggplot2 object.
}
\description{
Plot the evolution of arrival related times (flow, activity and waiting time).
}
\seealso{
\code{\link{plot_resource_usage}}, \code{\link{plot_resource_utilization}},
\code{\link{plot_attributes}}.
}

