% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_evolution_entity_times}
\alias{plot_evolution_entity_times}
\title{plot evolution of entity times}
\usage{
plot_evolution_entity_times(sim_obj, type = c("flow_time", "activity_time",
  "waiting_time"))
}
\arguments{
\item{sim_obj}{the simulation object}

\item{type}{one of c("flow_time","activity_time","waiting_time")}
}
\description{
plot the evolution of entity related times (flow, activity and waiting time)
}

