% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_resource_usage}
\alias{plot_resource_usage}
\title{plot usage of a resource over time}
\usage{
plot_resource_usage(simmer, resource_name, type = c("queue", "server",
  "system"), steps = FALSE)
}
\arguments{
\item{simmer}{the simulation environment}

\item{resource_name}{the name of the resource (character value)}

\item{type}{the parts of the resource to be plotted}

\item{steps}{adds the changes in the resource usage}
}
\description{
plot the usage of a resource over the simulation time frame
}

