% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_resource_usage}
\alias{plot_resource_usage}
\title{plot usage of a resource over time}
\usage{
plot_resource_usage(sim_obj, resource_name, replication_n = FALSE,
  steps = FALSE, smooth_line = FALSE)
}
\arguments{
\item{sim_obj}{the simulation object}

\item{resource_name}{the name of the resource (character value)}

\item{replication_n}{specify to plot only a specific replication (default=FALSE)}

\item{steps}{adds the changes in the resource usage}
}
\description{
plot the usage of a resource over the simulation time frame
}

