% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_resource_usage}
\alias{plot_resource_usage}
\title{Plot usage of a resource over time}
\usage{
plot_resource_usage(envs, resource_name, items = c("system", "queue",
  "server"), steps = FALSE)
}
\arguments{
\item{envs}{a single simmer environment or a list of environments representing several replications.}

\item{resource_name}{the name of the resource (character value).}

\item{items}{the components of the resource to be plotted.}

\item{steps}{adds the changes in the resource usage.}
}
\value{
Returns a ggplot2 object.
}
\description{
Plot the usage of a resource over the simulation time frame.
}
\seealso{
\code{\link{plot_resource_utilization}},
\code{\link{plot_evolution_arrival_times}}, \code{\link{plot_attributes}}.
}

