% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_resource_usage}
\alias{plot_resource_usage}
\title{plot usage of a resource over time}
\usage{
plot_resource_usage(reps, resource_name, items = c("queue", "server",
  "system"), steps = FALSE)
}
\arguments{
\item{reps}{a single simmer environment or a list of environments representing several replications}

\item{resource_name}{the name of the resource (character value)}

\item{items}{the components of the resource to be plotted}

\item{steps}{adds the changes in the resource usage}
}
\description{
plot the usage of a resource over the simulation time frame
}

