% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_resource_utilization}
\alias{plot_resource_utilization}
\title{plot utilization of resources}
\usage{
plot_resource_utilization(sim_obj)
}
\arguments{
\item{sim_obj}{the simulation object}
}
\description{
plot the utilization of all resources in the simulation
}

