% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_resource_utilization}
\alias{plot_resource_utilization}
\title{Plot utilization of resources}
\usage{
plot_resource_utilization(envs, resources)
}
\arguments{
\item{envs}{a single simmer environment or a list of environments representing several replications}

\item{resources}{a character vector with at least one resource specified - e.g. "c('res1','res2')"}
}
\description{
Plot the utilization of specified resources in the simulation.
}

