% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{release}
\alias{release}
\title{Add a release activity}
\usage{
release(traj, resource, amount = 1)
}
\arguments{
\item{traj}{the trajectory object.}

\item{resource}{the name of the resource.}

\item{amount}{the amount to release, accepts either a callable object (a function) or a numeric value.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of releasing a resource to the tail of a trajectory.
}
\seealso{
\code{\link{seize}}, \code{\link{seize_selected}}, 
\code{\link{release_selected}}, \code{\link{select}}.
}

