% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{release}
\alias{release}
\title{Add a release activity}
\usage{
release(traj, resource, amount = 1)
}
\arguments{
\item{traj}{the trajectory object.}

\item{resource}{the name of the resource.}

\item{amount}{the amount to release.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of releasing a resource to the tail of a trajectory.
}
\seealso{
Other methods for dealing with trajectories:
\link{create_trajectory}, \link{get_head},
\link{get_tail}, \link{get_n_activities}, \link{seize}, \link{timeout}, 
\link{set_attribute}, \link{branch}, \link{rollback}.
}

