% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{release_selected}
\alias{release_selected}
\title{Add a release activity}
\usage{
release_selected(traj, amount = 1)
}
\arguments{
\item{traj}{the trajectory object.}

\item{amount}{the amount to release, accepts either a callable object (a function) or a numeric value.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of releasing a previously selected resource to the tail of a trajectory.
}
\seealso{
\link{seize_selected}, \link{select}. 
Other methods for dealing with trajectories:
\link{create_trajectory}, \link{get_head}, \link{get_tail}, 
\link{get_n_activities}, \link{join}, \link{seize}, \link{release}, \link{timeout}, 
\link{set_attribute}, \link{branch}, \link{rollback}.
}

