% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{renege_in}
\alias{renege_abort}
\alias{renege_in}
\title{Add a renege activity}
\usage{
renege_in(traj, t, out = NULL)

renege_abort(traj)
}
\arguments{
\item{traj}{the trajectory object.}

\item{t}{timeout to trigger reneging, accepts either a numeric or a callable object
(a function) which must return a numeric.}

\item{out}{optional sub-trajectory in case of reneging.}
}
\value{
The trajectory object.
}
\description{
Set or unset a timer after which the arrival will abandon.
}

