% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{renege_in}
\alias{renege_in}
\alias{renege_if}
\alias{renege_abort}
\title{Renege on some Condition}
\usage{
renege_in(.trj, t, out = NULL)

renege_if(.trj, signal, out = NULL)

renege_abort(.trj)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{t}{timeout to trigger reneging, accepts either a numeric or a callable object
(a function) which must return a numeric.}

\item{out}{optional sub-trajectory in case of reneging.}

\item{signal}{signal to trigger reneging, accepts either a string or a callable object
(a function) which must return a string.}
}
\value{
Returns the trajectory object.
}
\description{
Activities for setting or unsetting a timer or a signal after which the arrival will abandon.
}
\seealso{
\code{\link{send}}, \code{\link{leave}}
}
