% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer-methods.R
\name{reset}
\alias{reset}
\title{Reset a Simulator}
\usage{
reset(.env)
}
\arguments{
\item{.env}{the simulation environment.}
}
\value{
Returns the simulation environment.
}
\description{
Reset the following components of a simulation environment:
time, event queue, resources, generators and statistics.
}
\seealso{
\code{\link{stepn}}, \code{\link{run}}.
}
