% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{rollback}
\alias{rollback}
\title{Add a rollback activity}
\usage{
rollback(traj, amount, times = 1)
}
\arguments{
\item{traj}{the trajectory object.}

\item{amount}{the amount of activities to roll back (of the same level; it does
not go into branches).}

\item{times}{the number of repetitions until an arrival may continue.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity that goes backwards in the trajectory.
}
\seealso{
Other methods to deal with trajectories:
\link{create_trajectory}, \link{show_trajectory}, \link{get_head},
\link{get_tail}, \link{get_n_activities}, \link{seize}, \link{release},
\link{timeout}, \link{branch}.
}

