% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{rollback}
\alias{rollback}
\title{Add a rollback activity}
\usage{
rollback(traj, amount, times = 1, check)
}
\arguments{
\item{traj}{the trajectory object.}

\item{amount}{the amount of activities (of the same or parent trajectories) to roll back.}

\item{times}{the number of repetitions until an arrival may continue.}

\item{check}{a callable object (a function) that must return a boolean. If
present, the \code{times} parameter is ignored, and the activity uses this
function to check whether the rollback must be done or not.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity that goes backwards in the trajectory.
}

