% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{rollback}
\alias{rollback}
\title{Add a rollback activity}
\usage{
rollback(.trj, amount, times = 1, check)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{amount}{the amount of activities (of the same or parent trajectories) to roll back.}

\item{times}{the number of repetitions until an arrival may continue.}

\item{check}{a callable object (a function) which must return a boolean. If
present, the \code{times} parameter is ignored, and the activity uses this
function to check whether the rollback must be done or not.}
}
\value{
Returns the trajectory object.
}
\description{
Go backwards to a previous point in the trajectory. Useful to implement loops.
}

