% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer-methods.R
\name{run}
\alias{run}
\alias{stepn}
\title{Run a Simulation}
\usage{
run(.env, until = Inf, progress = NULL, steps = 10)

stepn(.env, n = 1)
}
\arguments{
\item{.env}{the simulation environment.}

\item{until}{stop time.}

\item{progress}{optional callback to show the progress of the simulation. The
completed ratio is periodically passed as argument to the callback.}

\item{steps}{number of steps to show as progress (it takes effect only if
\code{progress} is provided).}

\item{n}{number of events to simulate.}
}
\value{
Returns the simulation environment.
}
\description{
Execute steps until a given criterion.
}
\seealso{
\code{\link{reset}}.
}
