% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator-methods.R
\name{run}
\alias{onestep}
\alias{run}
\title{Run the simulation}
\usage{
run(.env, until = 1000)

onestep(.env)
}
\arguments{
\item{.env}{the simulation environment.}

\item{until}{stop time.}
}
\value{
Returns the simulation environment.
}
\description{
Execute steps until the given criterion.
}
\seealso{
\code{\link{reset}}.
}

