% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{run}
\alias{run}
\title{Run the simulation}
\usage{
run(env, until = 1000)
}
\arguments{
\item{env}{the simulation environment.}

\item{until}{stop time.}
}
\value{
Returns the simulation environment.
}
\description{
Executes steps until the given criterion.
}
\seealso{
Other methods to deal with a simulation environment:
\link{simmer}, \link{reset}, \link{now}, \link{peek}, \link{onestep}, 
\link{add_resource}, \link{add_generator}, \link{get_mon_arrivals}, \link{get_mon_resources}, 
\link{get_capacity}, \link{get_queue_size}, \link{get_server_count}, \link{get_queue_count}.
}

