% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{seize}
\alias{release}
\alias{release_selected}
\alias{seize}
\alias{seize_selected}
\title{Add a seize/release activity}
\usage{
seize(traj, resource, amount = 1, ...)

seize_selected(traj, amount = 1, id = 0, ...)

release(traj, resource, amount = 1)

release_selected(traj, amount = 1, id = 0)
}
\arguments{
\item{traj}{the trajectory object.}

\item{resource}{the name of the resource.}

\item{amount}{the amount to seize/release, accepts either a callable object (a function) or a numeric value.}

\item{...}{unused arguments}

\item{id}{selection identifier for nested usage.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of seizing/releasing a resource to the tail of a trajectory.
}
\seealso{
\code{\link{select}}.
}

