% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{seize}
\alias{seize}
\title{Add a seize activity}
\usage{
seize(traj, resource, amount = 1, priority = 0, preemptible = 0,
  restart = FALSE)
}
\arguments{
\item{traj}{the trajectory object.}

\item{resource}{the name of the resource.}

\item{amount}{the amount to seize.}

\item{priority}{the priority of the seize (a higher integer equals higher 
priority; defaults to the minimum priority, which is 0).}

\item{preemptible}{if the seize occurs in a preemptive resource, this parameter 
establishes the minimum incoming priority that can preempt this arrival (a seize 
with a priority equal to `preemptible` or more gains the resource; by default, 
any seize may cause preemption in a preemptive resource).}

\item{restart}{whether the activity must be restarted after being preempted.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of seizing a resource to the tail of a trajectory.
}
\seealso{
Other methods to deal with trajectories:
\link{create_trajectory}, \link{get_head},
\link{get_tail}, \link{get_n_activities}, \link{release}, \link{timeout}, 
\link{set_attribute}, \link{branch}, \link{rollback}.
}

