% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{seize}
\alias{seize}
\title{Add a seize activity}
\usage{
seize(traj, resource, amount = 1)
}
\arguments{
\item{traj}{the trajectory object.}

\item{resource}{the name of the resource.}

\item{amount}{the amount to seize.}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of seizing a resource to the tail of a trajectory.
}
\seealso{
Other methods to deal with trajectories:
\link{create_trajectory}, \link{show_trajectory}, \link{get_head},
\link{get_tail}, \link{get_n_activities}, \link{release}, \link{timeout}, 
\link{branch}, \link{rollback}.
}

