% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{seize_selected}
\alias{seize_selected}
\title{Add a seize activity}
\usage{
seize_selected(traj, amount = 1, id = 0, ...)
}
\arguments{
\item{traj}{the trajectory object.}

\item{amount}{the amount to seize, accepts either a callable object (a function) or a numeric value.}

\item{id}{selection identifier for nested usage.}

\item{...}{unused arguments}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of seizing a previously selected resource to the tail of a trajectory.
}
\seealso{
\code{\link{seize}}, \code{\link{release}}, 
\code{\link{release_selected}}, \code{\link{select}}.
}

