% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{set_attribute}
\alias{set_attribute}
\alias{set_global}
\title{Set Attributes}
\usage{
set_attribute(.trj, key, value, global = FALSE)

set_global(.trj, key, value)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{key}{the attribute name.}

\item{value}{the value to set, accepts either a numeric or a callable object
(a function) which must return a numeric.}

\item{global}{if \code{TRUE}, the attribute will be global instead of per-arrival.}
}
\value{
Returns the trajectory object.
}
\description{
Activity for modifying an arrival's attribute in the form of a key/value pair.
}
\details{
\code{set_global} is a shortcut for \code{set_attribute(global=TRUE)}.
}
\seealso{
\code{\link{get_attribute}}.
}
