% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{set_attribute}
\alias{set_attribute}
\title{Add a set attribute activity}
\usage{
set_attribute(traj, key, value)
}
\arguments{
\item{traj}{the trajectory object.}

\item{key}{the attribute key (is coerced to a string).}

\item{value}{the value to set, accepts either a numeric or a callable object 
(a function) which returns a numeric.}
}
\value{
The trajectory object.
}
\description{
Adds or modifies a key/value attribute. The value must be numeric.
}

