% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{set_attribute}
\alias{set_attribute}
\alias{set_global}
\title{Set Attributes}
\usage{
set_attribute(.trj, keys, values, global = FALSE, mod = c(NA, "+", "*"))

set_global(.trj, keys, values, mod = c(NA, "+", "*"))
}
\arguments{
\item{.trj}{the trajectory object.}

\item{keys}{the attribute name(s), or a callable object (a function) which
must return attribute name(s).}

\item{values}{numeric value(s) to set, or a callable object (a function) which
must return numeric value(s).}

\item{global}{if \code{TRUE}, the attribute will be global instead of per-arrival.}

\item{mod}{if set, \code{values} modify the attributes rather than substituting them.}
}
\value{
Returns the trajectory object.
}
\description{
Activity for modifying an arrival's attributes.
}
\details{
\code{set_global} is a shortcut for \code{set_attribute(global=TRUE)}.
}
\seealso{
\code{\link{get_attribute}}.
}
