% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{set_attribute}
\alias{set_attribute}
\title{Add a set attribute activity}
\usage{
set_attribute(traj, key, value)
}
\arguments{
\item{traj}{the trajectory object.}

\item{key}{the attribute key (is coerced to a string).}

\item{value}{the value to set, accepts either a numeric or a callable object (a function) which returns a numeric.}
}
\value{
The trajectory object.
}
\description{
Adds a new key/value attribute. The value should be numeric.
}
\seealso{
Other methods for dealing with trajectories:
\link{create_trajectory}, \link{get_head}, \link{get_tail}, 
\link{get_n_activities}, \link{join}, \link{seize}, \link{release}, 
\link{timeout}, \link{branch}, \link{rollback},
\link{seize_selected}, \link{release_selected}, \link{select}.
}

