% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{set_capacity}
\alias{set_capacity}
\title{Set the capacity}
\usage{
set_capacity(env, name, value)
}
\arguments{
\item{env}{the simulation environment.}

\item{name}{the name of the resource.}

\item{value}{new capacity.}
}
\value{
Returns the simulation environment.
}
\description{
Sets a resource's capacity by name.
}
\seealso{
\code{\link{get_capacity}}, \code{\link{get_queue_size}}, \code{\link{set_queue_size}}, 
\code{\link{get_server_count}}, \code{\link{get_queue_count}}.
}

