% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{set_capacity}
\alias{set_capacity}
\alias{set_capacity_selected}
\alias{set_queue_size}
\alias{set_queue_size_selected}
\title{Set Resource Parameters}
\usage{
set_capacity(.trj, resource, value, mod = c(NA, "+", "*"))

set_capacity_selected(.trj, value, id = 0, mod = c(NA, "+", "*"))

set_queue_size(.trj, resource, value, mod = c(NA, "+", "*"))

set_queue_size_selected(.trj, value, id = 0, mod = c(NA, "+", "*"))
}
\arguments{
\item{.trj}{the trajectory object.}

\item{resource}{the name of the resource.}

\item{value}{new value to set.}

\item{mod}{if set, \code{values} modify the attributes rather than substituting them.}

\item{id}{selection identifier for nested usage.}
}
\value{
Returns the trajectory object.
}
\description{
Activities for modifying a resource's server capacity or queue size, by name
or a previously selected one.
}
\seealso{
\code{\link{select}}, \code{\link{seize}}, \code{\link{release}},
\code{\link{seize_selected}}, \code{\link{release_selected}},
\code{\link{get_capacity}}, \code{\link{get_queue_size}}.
}
