% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{set_capacity}
\alias{set_capacity}
\title{Set the capacity}
\usage{
set_capacity(env, name, value)
}
\arguments{
\item{env}{the simulation environment.}

\item{name}{the name of the resource.}

\item{value}{new capacity.}
}
\value{
Returns the simulation environment.
}
\description{
Sets a resource's capacity by name.
}
\seealso{
Other methods for dealing with a simulation environment:
\link{simmer}, \link{reset}, \link{now}, \link{peek}, \link{onestep}, \link{run}, 
\link{add_resource}, \link{add_generator}, \link{get_mon_arrivals}, \link{get_mon_attributes},
\link{get_mon_resources}, \link{get_n_generated}, \link{get_capacity}, \link{get_queue_size},
\link{set_queue_size}, \link{get_server_count}, \link{get_queue_count}.
}

