% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-methods.R
\name{set_capacity}
\alias{set_capacity}
\alias{set_capacity_selected}
\alias{set_queue_size}
\alias{set_queue_size_selected}
\title{Add a set capacity/queue size activity}
\usage{
set_capacity(.trj, resource, value)

set_capacity_selected(.trj, value, id = 0)

set_queue_size(.trj, resource, value)

set_queue_size_selected(.trj, value, id = 0)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{resource}{the name of the resource.}

\item{value}{new value to set.}

\item{id}{selection identifier for nested usage.}
}
\value{
Returns the trajectory object.
}
\description{
Modify a resource's server capacity or queue size, by name or a previously selected one.
}
\seealso{
\code{\link{select}}, \code{\link{seize}}, \code{\link{release}},
\code{\link{seize_selected}}, \code{\link{release_selected}}.
}

