% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{set_prioritization}
\alias{set_prioritization}
\title{Add a set prioritization activity}
\usage{
set_prioritization(traj, values)
}
\arguments{
\item{traj}{the trajectory object.}

\item{values}{expects either a vector/list or a callable object (a function)
returning a vector/list of three values \code{c(priority, preemptible, restart)}.
A negative value leaves the corresponding parameter unchanged.
See \code{\link{add_generator}} for more information about these parameters.}
}
\value{
The trajectory object.
}
\description{
Modifies the arrival's prioritization values.
}

