% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{set_trajectory}
\alias{set_trajectory}
\alias{set_source}
\alias{set_distribution}
\title{Set Source Parameters}
\usage{
set_trajectory(.trj, source, trajectory)

set_source(.trj, source, object)

set_distribution(.trj, source, object)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{source}{the name of the source or a function returning a name.}

\item{trajectory}{the trajectory that the generated arrivals will follow.}

\item{object}{a function modelling the interarrival times (if the source type
is a generator; returning a negative value stops the generator) or a data frame
(if the source type is a data source).}
}
\value{
Returns the trajectory object.
}
\description{
Activities for modifying a source's trajectory or source object by name.
}
\seealso{
\code{\link{activate}}, \code{\link{deactivate}}.
}
