% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simmer}
\alias{simmer}
\title{Start the simulation}
\usage{
simmer(sim_obj, until = Inf)
}
\arguments{
\item{sim_obj}{the Simulator object}

\item{until}{the maximum simulation time (after which the simulation is stopeed)}
}
\description{
Start the simulation
}

