% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulator.R
\docType{data}
\name{Simmer}
\alias{Simmer}
\title{Simmer}
\format{An \code{\link{R6Class}} generator object}
\usage{
Simmer
}
\description{
Simulation environment
}
\section{Fields}{

\describe{
\item{\code{name}}{environment name}
}}
\examples{
simmer <- Simmer$new("SuperDuperSim", rep=100, verbose=F) $
  add_resource("nurse", 1) $
  add_resource("doctor", 2) $
  add_resource("administration", 1) $
  add_generator("patient", t1, function() rnorm(1, 10, 2))
simmer$run(until=80)
}
\keyword{datasets}

