% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulator.R
\docType{data}
\name{Simmer}
\alias{Simmer}
\title{Simmer}
\format{An \code{\link{R6Class}} generator object}
\usage{
Simmer
}
\description{
Simulator object
}
\section{Fields}{

\describe{
\item{\code{name}}{simulator name}
}}
\keyword{datasets}

