% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap.R
\docType{data}
\name{Simmer.wrap}
\alias{Simmer.wrap}
\title{Simmer.wrap}
\description{
Extracts the simulation data from a Simmer object making it accessible
through the same methods. Only useful if you want to parallelize heavy
replicas (see the example below), because the C++ simulation backend is
destroyed when the threads exit.
}
\section{Methods}{

\preformatted{## Object creation
Simmer.wrap$new(simmer)
}\describe{
  \item{simmer}{the \link{Simmer} object}
}
\preformatted{## Get arrival statistics
Simmer.wrap$get_mon_arrivals()
}
\preformatted{## Get resource statistics
Simmer.wrap$get_mon_resources()
}
\preformatted{## Get resource's capacity
Simmer$get_capacity(name)
}\describe{
  \item{name}{the name of the resource}
}
\preformatted{## Get resource's queue size
Simmer$get_queue_size(name)
}\describe{
  \item{name}{the name of the resource}
}
}
\examples{
library(parallel)

mm1 <- Trajectory$new() $
  seize("server", 1) $
  timeout(function() rexp(1, 2)) $
  release("server", 1)

reps <- mclapply(1:4, function(i) {
  Simmer.wrap$new(
  Simmer$new("M/M/1 example") $
    add_resource("server", 1) $
    add_generator("customer", mm1, function() rexp(1, 1)) $
    run(100)
  )
})

plot_resource_usage(reps, "server")
}
\seealso{
\link{Simmer}
}
\keyword{datasets}

