% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-activities.R
\name{timeout}
\alias{timeout}
\alias{timeout_from_attribute}
\alias{timeout_from_global}
\title{Delay}
\usage{
timeout(.trj, task)

timeout_from_attribute(.trj, key)

timeout_from_global(.trj, key)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{task}{the timeout duration supplied by either passing a numeric or a
callable object (a function) which must return a numeric (negative values are
automatically coerced to positive).}

\item{key}{the attribute name, or a callable object (a function) which
must return the attribute name.}
}
\value{
Returns the trajectory object.
}
\description{
Activity for inserting delays and execute user-defined tasks.
}
\seealso{
\code{\link{set_attribute}}, \code{\link{set_global}}.
}
