% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{timeout}
\alias{timeout}
\title{Add a timeout activity}
\usage{
timeout(traj, task)
}
\arguments{
\item{traj}{the trajectory object.}

\item{task}{the timeout duration supplied by either passing a numeric value or a callable object (a function) that returns a numeric value 
(negative values are automatically coerced to positive).}
}
\value{
The trajectory object.
}
\description{
Adds a new activity capable of executing any user-defined task and setting
an associated delay to the tail of a trajectory.
}
\seealso{
Other methods for dealing with trajectories:
\link{create_trajectory}, \link{get_head},
\link{get_tail}, \link{get_n_activities}, \link{seize}, \link{release}, 
\link{set_attribute}, \link{branch}, \link{rollback}.
}

