% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{timeout}
\alias{timeout}
\title{Add a timeout activity}
\usage{
timeout(traj, task)
}
\arguments{
\item{traj}{the trajectory object.}

\item{task}{the timeout duration supplied by either passing a numeric or a
callable object (a function) which must return a numeric (negative values are
automatically coerced to positive).}
}
\value{
The trajectory object.
}
\description{
Insert delays and execute user-defined tasks.
}

