% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trajectory.R
\docType{data}
\name{Trajectory}
\alias{Trajectory}
\title{Trajectory}
\format{An \code{\link{R6Class}} generator object}
\usage{
Trajectory
}
\description{
Trajectory object
}
\section{Fields}{

\describe{
\item{\code{name}}{trajectory name}

\item{\code{...}}{events}
}}
\keyword{datasets}

