% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\docType{data}
\name{Trajectory}
\alias{Trajectory}
\title{Trajectory}
\description{
A trajectory is a chain of activities followed by arrivals of the same type.
}
\section{Methods}{

\preformatted{## Object creation
Trajectory$new(name="anonymous")
}\describe{
  \item{name}{the name of the trajectory}
}
\preformatted{## Show the trajectory
Trajectory$show() 
}\preformatted{## Get the first activity
Trajectory$get_head()
}
\preformatted{## Get the last activity
Trajectory$get_tail()
}
\preformatted{## Get the total number of activities in the trajectory
Trajectory$get_n_activities()
}
\preformatted{## Add a seize activity
Trajectory$seize(resource, amount)
}\describe{
  \item{resource}{the name of the resource}
  \item{amount}{the amount needed}
}
\preformatted{## Add a release activity
Trajectory$release(resource, amount)
}\describe{
  \item{resource}{the name of the resource}
  \item{amount}{the amount needed}
}
\preformatted{## Add a delay activity
Trajectory$timeout(duration)
}\describe{
  \item{duration}{the delay}
}
\preformatted{## Add n branches in the trajectory
Trajectory$branch(option, merge, ...)
}\describe{
  \item{option}{a function that must return a number between 1:n to select a branch}
  \item{merge}{a vector of n booleans that indicate whether the arrival must continue executing activities after each branch or not}
  \item{...}{n Trajectory objects describing each branch}
}
}
\examples{
t0 <- Trajectory$new("my trajectory") $
  ## add an intake activity
  seize("nurse", 1) $
  timeout(function() rnorm(1, 15)) $
  release("nurse", 1) $
  ## add a consultation activity
  seize("doctor", 1) $
  timeout(function() rnorm(1, 20)) $
  release("doctor", 1) $
  ## add a planning activity
  seize("administration", 1) $
  timeout(function() rnorm(1, 5)) $
  release("administration", 1)

t0$show()

t1 <- Trajectory$new("trajectory with a branch") $
  seize("server", 1) $
  ## 50-50 chance for each branch
  branch(function() sample(1:2, 1), merge=c(TRUE, FALSE), 
    Trajectory$new("branch1") $
      timeout(function() 1),
    Trajectory$new("branch2") $
      timeout(function() rexp(1, 3)) $
      release("server", 1)
  ) $
  ## only the first branch continues here
  release("server", 1) $
  timeout(function() 2)

t1$show()
}
\seealso{
\link{Simmer}
}
\keyword{datasets}

