% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trajectory.R
\docType{data}
\name{Trajectory}
\alias{Trajectory}
\title{Trajectory}
\format{An \code{\link{R6Class}} generator object}
\usage{
Trajectory
}
\description{
Trajectory object
}
\section{Fields}{

\describe{
\item{\code{name}}{trajectory name}
}}
\examples{
library(simmer)

t0 <- Trajectory$new("my trajectory") $
  ## add an intake activity
  seize("nurse", 1) $
  timeout(function() rnorm(1, 15)) $
  release("nurse", 1) $
  ## add a consultation activity
  seize("doctor", 1) $
  timeout(function() rnorm(1, 20)) $
  release("doctor", 1) $
  ## add a planning activity
  seize("administration", 1) $
  timeout(function() rnorm(1, 5)) $
  release("administration", 1)
}
\keyword{datasets}

